[Hotfix]  Handle edge case ALC LC off blinker on
Added `self.blinker_has_lane_change` to handle the edge case where, during an ALC lane change, the lane change is cancelled due to a blind spot and the speed drops below 50 km/h while the turn signal is still on. Lane keep should remain active because the driver might not be ready for the change.